organised string functions
